+++ src/core/event-name.enum.ts
@@ -50,9 +50,15 @@ export const enum EventName {
   applePayQrOpened = 'applePayQrOpened',
   applePayQrClosed = 'applePayQrClosed',
   applePayButtonClicked = 'applePayButtonClicked',
+  // Coupon
+  applyCoupon = 'applyCoupon',
+  removeCoupon = 'removeCoupon',
+  couponUpdate = 'couponUpdate',
   // Google Pay
   googlePayButtonClicked = 'googlePayButtonClicked',
   googlePayWindowOpened = 'googlePayWindowOpened',
   googlePayWindowClosed = 'googlePayWindowClosed',
   googlePayError = 'googlePayError',
+  // Form loading
+  formLoaded = 'formLoaded',
 }