+++ src/core/exception-handling/common-log-attributes/common-log-attributes.interface.ts
@@ -1,0 +1,15 @@
+export interface CommonLogAttributes {
+  token?: string;
+  userAgent?: string;
+  referrerUrl?: string;
+  viewportWidth?: number;
+  viewportHeight?: number;
+  isWebview?: boolean;
+  theme?: string;
+  topLevelDomain?: string;
+  sandbox?: boolean;
+  language?: string;
+  isApplePayInstantFlowEnabled?: boolean;
+  isGooglePayInstantFlowEnabled?: boolean | null;
+  apiUrl?: string;
+}