+++ src/core/exception-handling/elk/elk-error-types.enum.ts
@@ -1,0 +1,6 @@
+export enum ErrorTags {
+  PAY_STATION_SDK = 'payStationSDK',
+  ERROR = 'error',
+  INFO = 'info',
+  WARN = 'warn',
+}