+++ src/core/exception-handling/otel/services/otel-environments-tags.type.ts
@@ -1,0 +1,1 @@
+export type OtelEnvironmentTag = 'prod' | 'stage' | 'dev';