+++ src/core/guards/close-external-window.guard.ts
@@ -2,9 +2,15 @@ import { Message } from '../message.interface';
 import { isEventMessage } from './event-message.guard';
 import { EventName } from '../event-name.enum';
 
+export interface CloseExternalWindowMessageData {
+  closedByUser?: boolean;
+}
+
 export const isCloseExternalWindowMessage = (
   messageData: unknown,
-): messageData is Message<{ error: string } | null | undefined> => {
+): messageData is Message<
+  CloseExternalWindowMessageData | null | undefined
+> => {
   if (isEventMessage(messageData)) {
     return messageData.name === EventName.closeExternalWindow;
   }