+++ src/core/web-components/web-component-tag-name.enum.ts
@@ -29,4 +29,5 @@ export enum WebComponentTagName {
   QrCodeComponent = 'psdk-qr-code',
   CashPaymentInstructionComponent = 'psdk-cash-payment-instruction',
   XsollaNumberComponent = 'psdk-xsolla-number',
+  CouponComponent = 'psdk-coupon',
 }