+++ src/features/headless-checkout/web-components/payment-form/form-controls-tags.list.ts
@@ -6,4 +6,5 @@ export const formControlsTags = [
   WebComponentTagName.SelectComponent,
   WebComponentTagName.CheckboxComponent,
   WebComponentTagName.CardNumberComponent,
+  WebComponentTagName.CouponComponent,
 ];